+++ ncio.py
@@ -8,7 +8,7 @@ from math import sin, cos, sqrt, pi
 from .units import ang2bohr, degrad
 #from .atoms import atomic_weight, atomic_symbol, atomic_number
 from .atoms import *
-from .aux.convert import convert_abc2xyz, convert_xyz2abc, read_xyz
+from .aux import convert_abc2xyz, convert_xyz2abc, read_xyz
 
 def write_struct(atoms, cellparameter=1.0, fname = "STRUCT.fdf"):
